backend/requirements.txt
fastapi>=0.115.2
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.2
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.2
aioboto3>=12.4.0
tenacity>=8.4.1
supabase>=2.0.0
postgrest>=0.13.0
reportlab>=4.0.0
requests>=2.31.0
twilio>=8.10.0
psycopg2-binary>=2.9.9
easebuzz>=1.0.0

